requirements.txt
colorama==0.4.3
GitPython==3.1.27
python-slugify==6.1.2